parser and main

+++ src/main/scala/classes/Actor.scala
@@ -1,7 +1,7 @@
 package classes
 
 case class Actor  (
-                   id: Int,
+                   id: BigInt,
                    login: String,
                    display_login: String,
                    gravatar_id: String,
